ICMSLST-652 Rename web.domains.case._import.firearms to *.fa_oil
