+++ web/templates/forms/widgets/firearms/checkbox_select_table.html
@@ -23,7 +23,7 @@
               <td>
                 <a
                   href="{{ icms_url(
-                             'import:firearms:view-verified-certificate-file',
+                             'import:fa-oil:view-verified-certificate-file',
                              args=[process.pk, verified.pk, file.pk]) }}">
                   {{ file.filename }}
                 </a>
@@ -44,7 +44,7 @@
             {% elif user %}
               <td>
                 <a
-                  href="{{ icms_url('import:firearms:view-user-import-certificate-file', kwargs={'application_pk': process.pk, 'certificate_pk': file.pk}) }}">
+                  href="{{ icms_url('import:fa-oil:view-user-import-certificate-file', kwargs={'application_pk': process.pk, 'certificate_pk': file.pk}) }}">
                   {{ file.filename }}
                 </a>
               </td>
