+++ web/templates/partial/case/import/oil/sidebar.html
@@ -1,14 +1,14 @@
 <ul class="menu-out">
   <li>
-    <a href="{{ icms_url('import:firearms:edit-oil', kwargs={'pk': process.pk}) }}">Application Details</a>
+    <a href="{{ icms_url('import:fa-oil:edit-oil', kwargs={'pk': process.pk}) }}">Application Details</a>
   </li>
   <li>
-    <a href="{{ icms_url('import:firearms:list-user-import-certificates', kwargs={'pk': process.pk}) }}">Certificates</a>
+    <a href="{{ icms_url('import:fa-oil:list-user-import-certificates', kwargs={'pk': process.pk}) }}">Certificates</a>
   </li>
   <li>
-    <a href="{{ icms_url('import:firearms:list-import-contacts', kwargs={'pk': process.pk}) }}">Details of Who Bought From</a>
+    <a href="{{ icms_url('import:fa-oil:list-import-contacts', kwargs={'pk': process.pk}) }}">Details of Who Bought From</a>
   </li>
   <li>
-    <a href="{{ icms_url('import:firearms:submit-oil', kwargs={'pk': process.pk}) }}">Submit</a>
+    <a href="{{ icms_url('import:fa-oil:submit-oil', kwargs={'pk': process.pk}) }}">Submit</a>
   </li>
 </ul>
