+++ web/templates/partial/case/import/sidebar-management.html
@@ -10,13 +10,13 @@
   {% if process.process_type == "OpenIndividualLicenceApplication" %}
     <li>
       <a
-        href="{{ icms_url('import:firearms:manage-checklist', args=[process.pk]) }}">
+        href="{{ icms_url('import:fa-oil:manage-checklist', args=[process.pk]) }}">
         Checklist
       </a>
     </li>
     <li>
       <a
-        href="{{ icms_url('import:firearms:manage-constabulary-emails', args=[process.pk]) }}">
+        href="{{ icms_url('import:fa-oil:manage-constabulary-emails', args=[process.pk]) }}">
         Constabulary Emails
       </a>
     </li>
