+++ web/templates/partial/certificates/list.html
@@ -28,19 +28,19 @@
         <td>{{ certificate.expiry_date.strftime('%d-%b-%Y') }}</td>
         <td>
             <a
-              href="{{ icms_url('import:firearms:view-user-import-certificate-file', kwargs={'application_pk': process.pk, 'certificate_pk': certificate.pk}) }}">
+              href="{{ icms_url('import:fa-oil:view-user-import-certificate-file', kwargs={'application_pk': process.pk, 'certificate_pk': certificate.pk}) }}">
               {{ certificate.filename }}
             </a> {{ certificate.human_readable_file_size() }}
         </td>
         {% if not read_only %}
           <td>
             <form method="post"
-               action="{{ icms_url('import:firearms:delete-user-import-certificate', kwargs={'application_pk': process.pk, 'certificate_pk': certificate.pk}) }}">
+               action="{{ icms_url('import:fa-oil:delete-user-import-certificate', kwargs={'application_pk': process.pk, 'certificate_pk': certificate.pk}) }}">
              {{ csrf_input }}
              <button type="submit" class="link-button icon-bin button">Delete</button>
             </form>
             <p>
-              <a href="{{ icms_url('import:firearms:edit-user-import-certificate', kwargs={'application_pk': process.pk, 'certificate_pk': certificate.pk}) }}"
+              <a href="{{ icms_url('import:fa-oil:edit-user-import-certificate', kwargs={'application_pk': process.pk, 'certificate_pk': certificate.pk}) }}"
                  class="link-button icon-pencil button">
                 Edit</a>
             </p>
@@ -57,7 +57,7 @@
     <ul class="menu-out flow-across">
       <li>
         <a
-          href="{{ icms_url('import:firearms:create-user-import-certificate', kwargs={'pk': process.pk}) }}"
+          href="{{ icms_url('import:fa-oil:create-user-import-certificate', kwargs={'pk': process.pk}) }}"
           class="button small-button icon-plus">
         Add Certificate
         </a>
