+++ web/templates/web/domains/case/import/firearms/certificates/create.html
@@ -11,7 +11,7 @@
   <ul class="menu-out flow-across">
     <li>
       <a
-        href="{{ icms_url('import:firearms:list-user-import-certificates', kwargs={'pk': process.pk}) }}"
+        href="{{ icms_url('import:fa-oil:list-user-import-certificates', kwargs={'pk': process.pk}) }}"
         class="prev-link">
         Certificates
       </a>
