+++ web/templates/web/domains/case/import/firearms/edit-constabulary-email.html
@@ -19,7 +19,7 @@
   <div class="content-actions">
     <ul class="menu-out flow-across">
       <li>
-        <a href="{{ icms_url('import:firearms:manage-constabulary-emails', args=[process.pk]) }}" class="prev-link">
+        <a href="{{ icms_url('import:fa-oil:manage-constabulary-emails', args=[process.pk]) }}" class="prev-link">
           Constabulary Emails
         </a>
       </li>
