+++ web/templates/web/domains/case/import/firearms/import-contacts/create.html
@@ -12,7 +12,7 @@
   <ul class="menu-out flow-across">
     <li>
       <a
-        href="{{ icms_url('import:firearms:list-import-contacts', kwargs={'pk': process.pk}) }}"
+        href="{{ icms_url('import:fa-oil:list-import-contacts', kwargs={'pk': process.pk}) }}"
         class="prev-link">
         Details of Who Bought From
       </a>
