+++ web/templates/web/domains/case/import/firearms/import-contacts/list.html
@@ -10,7 +10,7 @@
   <ul class="menu-out flow-across">
     <li>
       <a
-        href="{{ icms_url('import:firearms:edit-oil', kwargs={'pk': process.pk}) }}"
+        href="{{ icms_url('import:fa-oil:edit-oil', kwargs={'pk': process.pk}) }}"
         class="prev-link">
         Edit Open Individual Licence
         {% if process.applicant_reference %}
