+++ web/tests/domains/case/_import/fa_oil/test_views.py
@@ -3,7 +3,7 @@ from django.test import Client
 from django.urls import reverse
 from guardian.shortcuts import assign_perm
 
-from web.domains.case._import.firearms.models import OpenIndividualLicenceApplication
+from web.domains.case._import.fa_oil.models import OpenIndividualLicenceApplication
 from web.domains.case._import.models import ImportApplicationType
 from web.domains.importer.models import Importer
 from web.tests.auth import AuthTestCase
