+++ web/tests/domains/case/_import/factory.py
@@ -1,7 +1,7 @@
 import factory.fuzzy
 
 from web.domains.case._import.derogations.models import DerogationsApplication
-from web.domains.case._import.firearms.models import OpenIndividualLicenceApplication
+from web.domains.case._import.fa_oil.models import OpenIndividualLicenceApplication
 from web.domains.case._import.models import ImportApplicationType
 from web.domains.case._import.sanctions.models import (
     SanctionsAndAdhocApplication,
